% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineclass.R
\docType{class}
\name{gcprofile-class}
\alias{gcprofile-class}
\title{Class gcprofile}
\description{
This is a class representation grouping different outputs from \code{getGroups} function
}
\section{Slots}{

\describe{
\item{\code{data}}{Data frame from \code{NormalizeBetweenFiles} function}

\item{\code{profile}}{Data frame with compounds classified as "Profile".
List of CAS numbers associated to a list of proposed names.}

\item{\code{nonConstitutiveFreq}}{Data frame with compounds classified as "Non Constitutive by Frequency".
List of CAS numbers associated to a list of proposed names.}

\item{\code{nonConstitutiveQual}}{Data frame with compounds classified as "Non Constitutive by Quality".
List of CAS numbers associated to a list of proposed names.}

\item{\code{areas}}{Data frame with CAS and areas of each CAS}

\item{\code{quals}}{Data frame with CAS and qualities of each CAS}

\item{\code{AuxTable}}{Data frame with CAS numbers associated to their areas and qualities in each file.
Data presented is the result of the consensus reached by NormalizeWithinFiles and NormalizeBetweenFiles.}
}}

\seealso{
Examples in \code{\link{getGroups}}
}
\author{
Fernando Pérez-Sanz \code{fernando.perez8@um.es}, Victoria Ruiz \code{victoria.ruiz@upct.es1}
}
