% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{getGroups}
\usage{
getGroups(data, savefiles = FALSE, verbose = TRUE, qcutoff = 85,
  ncFreqCutoff = 0.3, pFreqCutoff = 0.9)
}
\arguments{
\item{data}{Object class normdata from NormalizeBetweenFiles() function}

\item{savefiles}{A boolean indicating whether to save the results in the working directory. Default FALSE}

\item{verbose}{A boolean indicating whether to show the profiles in console. Default FALSE}

\item{qcutoff}{A numerical integer value indicating the average quality cut-off point to determine which compounds should be
listed as "Profile" and "Non Constitutive by Quality". Default value = 85}

\item{ncFreqCutoff}{A numerical float value between 0-1 indicating minimal frequency that establishes compounds
considered as non constitutive by frequency. This value always must be lower than pFreqCutoff.
Default value = 0.3}

\item{pFreqCutoff}{A numerical float value between 0-1 indicating minimal frequency that establishes compounds
considered as profile. This value always must be higher than ncFreqCutoff. Default value = 0.9}
}
\value{
A gcprofile class object
}
\description{
A function to finally determine emission profiles and non-constitutive compounds emitted.
}
\examples{
## This example run the function with default parameters. 
## Where out2 is a normdata class object output from the NormalizeBetweenFiles function.
path <- paste(system.file(package = "gcProfileMakeR"), "/extdata", sep="")
cat(path)
out1 <- NormalizeWithinFiles(path = path)
out2 <- NormalizeBetweenFiles(data = out1)
out <- getGroups(data = out2)

## full param configuration
out <- getGroups(data = out2, savefiles = FALSE, verbose = TRUE, 
                   qcutoff = c(85), ncFreqCutoff = c(0.3), pFreqCutoff = c(0.9) )

}
\seealso{
\code{\link{NormalizeBetweenFiles}}
}
\author{
Fernando Pérez-Sanz \code{fernando.perez8@um.es}, Victoria Ruiz \code{victoria.ruiz@upct.es}
}
