% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeBetweenFiles.R
\name{NormalizeBetweenFiles}
\alias{NormalizeBetweenFiles}
\title{NormalizeBetweenFiles}
\usage{
NormalizeBetweenFiles(data, savefiles = FALSE, thr = 2,
  filterLowQual = 0)
}
\arguments{
\item{data}{Object class normdata from NormalizeWithinFiles() function}

\item{savefiles}{A boolean. If you want to save the resulting files on disk. Default: FALSE}

\item{thr}{Threshold of time to be considered between all files in order to unify compounds. Default 2}

\item{filterLowQual}{Numeric. Filters out hits with quality below the established parameter.
Values in range 0-100. Default: 0, means do not filter.}
}
\value{
A normdata class object
}
\description{
The function groups together compounds with similar retention time among all files,
 determining for each group which CAS will be the most representative for that group.
}
\examples{

## This example runs the function with default parameters. 
## Where out1 is a normdata class object output from the NormalizeWithinFiles function.
path <- paste(system.file(package = "gcProfileMakeR"), "/extdata", sep="")
out1 <- NormalizeWithinFiles(path = path)
out2 <- NormalizeBetweenFiles(data = out1)

## full param configuration
out2 <- NormalizeBetweenFiles(data = out1, thr = 2, savefiles = FALSE, filterLowQual = 0)

}
\seealso{
\code{\link{NormalizeWithinFiles}}
}
\author{
Fernando Pérez-Sanz \code{fernando.perez8@um.es}, Victoria Ruiz \code{victoria.ruiz@upct.es}
}
