% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeWithinFiles.R
\name{NormalizeWithinFiles}
\alias{NormalizeWithinFiles}
\title{NormalizeWithinFiles}
\usage{
NormalizeWithinFiles(path, type = "xls", thr = 0.2,
  savefiles = FALSE, filterLowQual = 0.99, cas2rm = NULL,
  minQuality = NULL)
}
\arguments{
\item{path}{Complete directory path where the files are located.}

\item{type}{Either 'xls' or 'csv'. Input file type. Default: xls}

\item{thr}{Establishes the threshold of retention time to be considered in each file
for each peak in order to unify compounds. Default 0.2}

\item{savefiles}{A boolean. If you want to save the resulting files on disk. Default: FALSE}

\item{filterLowQual}{Numeric. Filters out secondary hits with low 
quality comparing to the first hit in a particular retention time. 
Percentage of quality similarity of secondary hits with first hit for 
a given retention time. Values in range 0-1. filterLowQual = 0, means do not filter. 
Default: 0.99, means remove hits with lower quality than 99% 
of the first hit (with the highest quality for that retention time).}

\item{cas2rm}{A Vector or NULL. Character vector containing the CAS to be removed
from the analysis. Default: NULL means do not filter out.}

\item{minQuality.}{Numeric or NULL. Filter out hits with quality lower than minQuality value.}
}
\value{
A normdata class object
}
\description{
Reads and pre-processes xls and csv files, filling in blank fields
and grouping compounds with similar retention time and CAS numbers in common.
}
\examples{

## 

path <- paste(system.file(package = "gcProfileMakeR"), "/extdata", sep="")
out1 <- NormalizeWithinFiles(path = path)

## or with custom parameters
path <- paste(system.file(package = "gcProfileMakeR"), "/extdata", sep="")
out1 <- NormalizeWithinFiles(path = path, type = 'xls', thr = 0.2, savefile = FALSE,
                          filterLowQual = 0.99)


}
\author{
Fernando Pérez-Sanz \code{fernando.perez8@um.es}, Victoria Ruiz \code{victoria.ruiz@upct.es}
}
