% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroup.R
\name{plotGroup}
\alias{plotGroup}
\title{plotGroup}
\usage{
plotGroup(data, compoundType = "profile", maxBars = NA)
}
\arguments{
\item{data}{Output object from getGroups() function}

\item{compoundType}{Set of compounds to be plotted. Valid values are: "profile", "nonconstitutivebyfreq"
and "nonconstitutivebyqual". Its abbreviations "p", "ncf", "ncq" are also valid. Default value is "profile"}

\item{maxBars}{A numerical integer value indicating the maximum number of bars to be displayed. By default displays all bars}
}
\value{
ggplot object
}
\description{
The function plots both line chart with average qualities and bars chart with average areas of
 compounds grouped as “profile”, “non constititive by frequency” or “non constitutive by quality”
}
\examples{

## Where out is a gcprofile class object output from the getGroups function.
path <- paste(system.file(package = "gcProfileMakeR"), "/extdata", sep="")
out1 <- NormalizeWithinFiles(path = path)
out2 <- NormalizeBetweenFiles(data = out1)
out <- getGroups(data = out2)
p <- plotGroup(data = out, compoundType = "p")
p
## or directly
plotGroup(data = out, compoundType = "p")


}
\author{
Fernando Pérez-Sanz \code{fernando.perez8@um.es}, Victoria Ruiz \code{victoria.ruiz@upct.es}
}
